Extract profile approval authorizer
